🔨 refactor(app): 상품관리 · 쿠폰관리 컴포넌트 분리
- 상품관리 영역을 ProductManager 컴포넌트로 분리
- 쿠폰관리 영역을 CouponManager 컴포넌트로 분리
