+++ src/basic/types/product.ts
@@ -4,3 +4,5 @@ export interface ProductWithUI extends Product {
   description?: string;
   isRecommended?: boolean;
 }
+
+export type NewProductForm = Omit<ProductWithUI, "id"> & { id?: string };
